refactor to remove duplicate code
