+++ python/auto/parseBandS.py
@@ -137,37 +137,9 @@ class parseBandS:
     def getLabels(self):
         return self.diagram.getLabels()
 
-def pointtest7(a,b):
-    if "TY name" not in a:
-        raise AUTOExceptions.AUTORegressionError("No TY name label")
-    if "TY number" not in a:
-        raise AUTOExceptions.AUTORegressionError("No TY number label")
-    if "BR" not in a:
-        raise AUTOExceptions.AUTORegressionError("No BR label")
-    if "data" not in a:
-        raise AUTOExceptions.AUTORegressionError("No data label")
-    if "PT" not in a:
-        raise AUTOExceptions.AUTORegressionError("No PT label")
-    if "LAB" not in a:
-        raise AUTOExceptions.AUTORegressionError("No LAB label")
-    if len(a["data"]) != len(b["data"]):
-        raise AUTOExceptions.AUTORegressionError("Data sections have different lengths")
-   
-    
-def pointtest8(a,b):
-    keys = ['Type number', 'Type name',
-            'Free Parameters',  'Branch number', 'Parameter NULL vector',
-            'data', 'NCOL', 'Label', 'ISW', 'NTST',
-            'Point number', 'Parameters']
 
-    scratch=a['Parameters']
-    scratch=b['Parameters']
-    for key in keys:
-        if key not in a:
-            raise AUTOExceptions.AUTORegressionError("No %s label"%(key,))
-    if len(a["data"]) != len(b["data"]):
-        raise AUTOExceptions.AUTORegressionError("Data sections have different lengths")
-
+import pointtest7
+import pointtest8
 
 def test():
     foo = parseBandS()
