+++ python/auto/pointtest8.py
@@ -1,0 +1,16 @@
+from auto import AUTOExceptions
+    
+def pointtest8(a,b):
+    keys = ['Type number', 'Type name',
+            'Free Parameters',  'Branch number', 'Parameter NULL vector',
+            'data', 'NCOL', 'Label', 'ISW', 'NTST',
+            'Point number', 'Parameters']
+
+    scratch=a['Parameters']
+    scratch=b['Parameters']
+    for key in keys:
+        if key not in a:
+            raise AUTOExceptions.AUTORegressionError("No %s label"%(key,))
+    if len(a["data"]) != len(b["data"]):
+        raise AUTOExceptions.AUTORegressionError("Data sections have different lengths")
+
